+++ src/App.js
@@ -2,16 +2,16 @@ import React, { useEffect, useState } from 'react';
 import './App.css';
 import Navbar from './Components/Navbar/Navbar';
 import { BrowserRouter as Router, Route, Routes, Link } from 'react-router-dom';
-import AddMember from './Pages/MembersDashboard/AddMember';
-import MembersDashboard from './Pages/MembersDashboard/MembersDashboard';
-import Member from './Pages/MembersDashboard/Member';
+import AddMember from './Pages/Members/AddMember';
+import MembersHome from './Pages/Members/MembersHome';
+import Member from './Pages/Members/Member';
 import Reporting from './Pages/Reporting/Reporting';
 import SalesReport from './Pages/Reporting/SalesReport';
 import InventoryReport from './Pages/Reporting/InventoryReport';
 import supabase from "./Supabase/supabaseClient";
 import { Auth } from "@supabase/auth-ui-react";
 import { ThemeSupa } from "@supabase/auth-ui-shared";
-import Products from './Pages/Products/Products';
+import ProductsHome from './Pages/Products/ProductsHome';
 import AddProduct from './Pages/Products/AddProduct';
 
 function App() {
@@ -44,13 +44,13 @@ function App() {
                     <Navbar />
                     <Routes>
                         <Route path="/" exact element={<HomePage />} />
-                        <Route path="/members-dashboard" element={<MembersDashboard />} />
+                        <Route path="/members-home" element={<MembersHome />} />
                         <Route path="/member" element={<Member />} />
                         <Route path="/add-member" element={<AddMember />} />
                         <Route path="/reporting" element={<Reporting />} />
                         <Route path="/sales-report" element={<SalesReport />} />
                         <Route path="/inventory-report" element={<InventoryReport />} />
-                        <Route path="/products" element={<Products />} />
+                        <Route path="/products-home" element={<ProductsHome />} />
                         <Route path="/add-product" element={<AddProduct />} />
                     </Routes>
                     <footer>
@@ -74,8 +74,8 @@ const HomePage = () => {
                     facing challenges in meeting its members' expectations and satisfying their diverse grocery needs.
                     We're excited to introduce our new digital Members Record Management System.
                 </p>
-                <Link className="main-button dashboard-button" to="/members-dashboard">Go to Members Dashboard</Link>
-                <Link className="main-button dashboard-button" to="/products">Go to Products Home</Link>
+                <Link className="main-button dashboard-button" to="/members-home">Go to Members Home</Link>
+                <Link className="main-button dashboard-button" to="/products-home">Go to Products Home</Link>
                 <Link className="main-button reporting-button" to="/reporting">Go to Reporting</Link>
             </main>
         </div>
